4099 경고 무시
링커 경고인 4099 경고 무시
현재 필요 없는 기능임

+++ src/D3D/DemoProcess.cpp
@@ -132,7 +132,7 @@ LRESULT CALLBACK DemoProcess::WndProc(HWND hWnd, UINT message,
 		staticManagers->keyManager->OnMouseRightUp(LOWORD(lParam), HIWORD(lParam));
 		break;
 	case WM_MOUSEMOVE:
-		staticManagers->keyManager->OnMouseMove(wParam, LOWORD(lParam), HIWORD(lParam));
+		staticManagers->keyManager->OnMouseMove(static_cast<int>(wParam), LOWORD(lParam), HIWORD(lParam));
 		break;
 	default:
 		return DefWindowProc(hWnd, message, wParam, lParam);
